use unsigned int to store line numbers in GenomeNode (save space)

+++ src/libgtext/genome_feature.h
@@ -35,7 +35,7 @@ typedef void (*AttributeIterFunc)(const char *attr_name, const char *attr_value,
 const GenomeNodeClass* genome_feature_class(void);
 GenomeNode*            genome_feature_new(GenomeFeatureType*, Range, Strand,
                                           Str *filename,
-                                          unsigned long line_number);
+                                          unsigned int line_number);
 /* return the ``standard gene'' (mainly for testing purposes) */
 GenomeNode*            genome_feature_new_standard_gene(FeatureTypeFactory*);
 const char*            genome_feature_get_source(GenomeFeature*);
